refactor: moved calendar screen in screens folder
