+++ lib/screens/home_screen.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
 import 'package:provider/provider.dart';
-import 'calendar/calendar_screen.dart';
+import 'calendar_screen.dart';
 import '../model/absences.dart';
 import 'absence_screen.dart';
 import '../services/User_service.dart';
